Creating front end for product list

+++ packages/server/src/routers/user/post.user.js
@@ -225,12 +225,16 @@ router.post('/upload', auth, uploadUser.single('gambar'), async (req, res) => {
 
     const paths = postPath + dataValues.image;
 
-    fs.unlink(paths, (err) => {
-      if (err) {
-        console.error(err);
-        return;
-      }
-    });
+    console.log(dataValues.image.split('/')[3] != 'default-avatar.png');
+
+    if (dataValues.image.split('/')[3] != 'default-avatar.png') {
+      fs.unlink(paths, (err) => {
+        if (err) {
+          console.error(err);
+          return;
+        }
+      });
+    }
 
     return res.send({
       status: 'Success',
